+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added "No result" cells to FlowsheetDataGrid.
+
 ## 0.8.1 - (October 17, 2023)
 
 * Fixed
@@ -14,7 +17,6 @@
 
 * Added
   * Added base FlowsheetDataGrid component.
-  * Added "No result" cells to FlowsheetDataGrid.
 
 * Changed
   * Updated prop descriptions for better consistency and accuracy in the `terra-data-grid` component.